Create data for Emoji
